TypeCast double and float according to your Locale if default cast fails
In some regions decimal types have comma instead dot.  Currently spark-csv cast to Double or Float using .toDouble or toFloat scala functions.

If we have double or float types with comma (and the delimiter is another character) we could try to use the java.text.NumberFormat and get the default Locale.

Author: gasparms <gmunoz@stratio.com>

Closes #154 from gasparms/master.