Nullable quote character
This pull request adds to spark-csv the functionality to provide a null for the quote character, in the event that no quoting is desired. This is necessary to interoperate with libraries such as Scalding which use unquoted formats by default.

Since the uniVocity parser uses a non-nullable char for the quoting character, a null byte is substituted in CsvRelation.scala if null was provided and uniVocity used, just as is already done for commentChar.

A test is also added using a comma-separated file with unbalanced double-quote characters.

Author: James Blau <james@tresata.com>

Closes #89 from tresata-opensource/feat-nullable-quotation-character.