Fix for #308 write using specified dateFormat
Added support for writing to file with dateFormat that is specified in the options.  Added some basic tests for it.

I'm continuing to use SimpleDateFormat since that is what is used for reading. I would prefer joda's DateTimeFormat, but that can be a separate issue.

It's too bad CSVFormat does not allow specification of a dateFormat as it does for the null value. That would make this much easier.

Author: barrybecker <barrybecker@sgi.com>

Closes #310 from barrybecker4/master.