fix NPE that can occur for null dates
I know #310 was recently merged and closed, but I just made an additional change to fix a problem I noted when null dates are in the data. A check for null needs to be done. Let me know if I need to create a separate issue or if you can just merge this.

Author: barrybecker <barrybecker@sgi.com>
Author: Barry Becker <BBE@esi-group.com>

Closes #360 from barrybecker4/master.